feat: Use CozyDialog instead of Modal

+++ src/drive/web/modules/trash/components/DestroyConfirm.jsx
@@ -1,8 +1,11 @@
 import React from 'react'
 import classNames from 'classnames'
+
 import { useClient } from 'cozy-client'
-import Modal from 'cozy-ui/transpiled/react/Modal'
+import { ConfirmDialog } from 'cozy-ui/transpiled/react/CozyDialogs'
+import Button from 'cozy-ui/transpiled/react/Button'
 import { translate } from 'cozy-ui/transpiled/react/I18n'
+
 import { deleteFilesPermanently } from 'drive/web/modules/actions/utils'
 import styles from 'drive/styles/confirms.styl'
 
@@ -16,26 +19,34 @@ const DestroyConfirm = ({ t, files, onClose }) => {
       {t(`destroyconfirmation.${type}`, files.length)}
     </p>
   ))
-
   return (
-    <Modal
+    <ConfirmDialog
+      open={true}
+      onClose={onClose}
       title={t('destroyconfirmation.title', files.length)}
-      description={confirmationTexts}
-      secondaryType="secondary"
-      secondaryText={t('destroyconfirmation.cancel')}
-      dismissAction={onClose}
-      secondaryAction={onClose}
-      primaryType="danger"
-      primaryText={t('destroyconfirmation.delete')}
-      primaryAction={async () => {
-        try {
-          await deleteFilesPermanently(client, files)
-        } catch {
-          //eslint-disable-next-line
-        } finally {
-          onClose()
-        }
-      }}
+      content={confirmationTexts}
+      actions={
+        <>
+          <Button
+            theme="secondary"
+            onClick={onClose}
+            label={t('destroyconfirmation.cancel')}
+          />
+          <Button
+            theme="danger"
+            label={t('destroyconfirmation.delete')}
+            onClick={async () => {
+              try {
+                await deleteFilesPermanently(client, files)
+              } catch {
+                //eslint-disable-next-line
+            } finally {
+                onClose()
+              }
+            }}
+          />
+        </>
+      }
     />
   )
 }
